add messages and add authorization on bought

+++ src/pages/api/boughts/[boughtID].ts
@@ -7,33 +7,38 @@ export default async function bought(
   res: NextApiResponse
 ) {
   const { method } = req;
+  const { authorization } = req.headers;
   const { boughtID } = Array.isArray(req.query) ? req.query[0] : req.query;
   const { idItem, idUser, status, qtd } = req.body;
 
-  switch (method) {
-    case "PATCH":
-      const updateBought = await prisma.boughts.update({
-        where: {
-          id: boughtID,
-        },
-        data: {
-          status,
-          idItem,
-          idUser,
-          qtd,
-        },
-      });
-      return res.status(201).json(updateBought);
+  if (!authorization) {
+    return res.status(501).json({ message: "Unauthorized" });
+  } else {
+    switch (method) {
+      case "PATCH":
+        const updateBought = await prisma.boughts.update({
+          where: {
+            id: boughtID,
+          },
+          data: {
+            status,
+            idItem,
+            idUser,
+            qtd,
+          },
+        });
+        return res.status(201).json(updateBought);
 
-    case "DELETE":
-      const deletedBought = await prisma.boughts.delete({
-        where: {
-          id: boughtID,
-        },
-      });
-      return res.status(204).json(deletedBought);
+      case "DELETE":
+        const deletedBought = await prisma.boughts.delete({
+          where: {
+            id: boughtID,
+          },
+        });
+        return res.status(204).json(deletedBought);
 
-    default:
-      return res.status(404).json({ message: "Wrong method!" });
+      default:
+        return res.status(404).json({ message: "Wrong method!" });
+    }
   }
 }
